requirements.txt
colorama==0.4.6
pillow==11.3.0
pyfiglet==1.0.4
requests==2.32.5
rich==14.2.0
yt-dlp==2025.10.22
